+++ app/src/main/java/com/mapbox/maps/testapp/examples/customlayer/CustomLayerActivity.kt
@@ -56,7 +56,7 @@ class CustomLayerActivity : AppCompatActivity() {
   }
 
   private fun swapCustomLayer() {
-    mapboxMap.getStyle()?.let { style ->
+    mapboxMap.style?.let { style ->
       if (style.styleLayerExists(CUSTOM_LAYER_ID)) {
         style.removeStyleLayer(CUSTOM_LAYER_ID)
         binding.fab.setImageResource(R.drawable.ic_layers)