+++ extension-compose/src/main/java/com/mapbox/maps/extension/compose/animation/viewport/MapViewportState.kt
@@ -141,7 +141,7 @@ public class MapViewportState(
    */
   @MapboxExperimental
   public val styleDefaultCameraOptions: CameraOptions?
-    get() = controller?.getMapboxMap()?.getStyle()?.styleDefaultCamera
+    get() = controller?.getMapboxMap()?.style?.styleDefaultCamera
 
   /**
    * Ease the map camera to a given camera options.